requirements.txt
flask==2.0.1
flask-cors==3.0.10
firebase-admin==5.0.4
python-dotenv==0.19.0 